+++ src/readtags.js
@@ -15,7 +15,7 @@ class ReadtagsProvider {
         const command = getConfiguration(scope).get("readtagsGoToDefinitionCommand");
         const cwd = scope.uri.fsPath;
 
-        const definitions = await this.exec(`${command} ${symbol}`, { cwd });
+        const definitions = await this.exec(`${command} '${symbol}'`, { cwd });
         return definitions
             .map(d => definitionToSymbolInformation(d, scope))
             .map(({ location }) => location);
@@ -27,7 +27,7 @@ class ReadtagsProvider {
         const results = await Promise.all(vscode.workspace.workspaceFolders.map(async scope => {
             const command = getConfiguration(scope).get("readtagsGoToSymbolInWorkspaceCommand");
             const cwd = scope.uri.fsPath;
-            const definitions = await this.exec(`${command} ${query}`, { cwd, maxBuffer: Infinity });
+            const definitions = await this.exec(`${command} '${query}'`, { cwd, maxBuffer: Infinity });
             return definitions.map(d => definitionToSymbolInformation(d, scope));
         }));
         return results.flat();
@@ -39,7 +39,7 @@ class ReadtagsProvider {
         const relativePath = vscode.workspace.asRelativePath(document.uri, false);
         const cwd = scope.uri.fsPath;
 
-        const definitions = await this.exec(`${command} ${relativePath}`, { cwd });
+        const definitions = await this.exec(`${command} '${relativePath}'`, { cwd });
         return definitions.map(d => definitionToSymbolInformation(d, scope));
     }
 }